🗺️ Atlas: [parser structure] Extract `build_statement` to break cycle
🕸️ Tangle: `src/parser/mod.rs` exported `build_statement` which was imported by its children (`expressions.rs`, `declarations.rs`), creating a cyclic dependency graph between the parent and child modules.
📐 Blueprint: Extracted `build_statement` implementation to `src/parser/statements.rs` where statement parsing logically belongs, and updated `mod.rs` and sibling child modules to import it from there.
🧱 Stability: Broke the circular dependency, restoring a strict unidirectional graph. Reduced the size of `mod.rs`.
🔬 Verification: Ran `cargo check` and `cargo test`, ensuring all module paths resolved without cyclic constraints, and `cargo clippy` passed cleanly.
