+++ src/parser/expressions.rs
@@ -2,9 +2,9 @@
 //!
 //! Builds AST expressions from parsing grammar pairs.
 use crate::ast::*;
-use crate::parser::build_statement;
 use crate::parser::common::{ParseError, parse_number_literal};
 use crate::parser::grammar::Rule;
+use crate::parser::statements::build_statement;
 use pest::iterators::Pair;
 
 /// Builds an expression from a grammar pair.
